# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Budget App API
  version: 1.0.0
  description: REST API for the budget tracking application
servers:
  - url: http://localhost:5000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CategoryTemplate:
      type: object
      required: [id, name, color, defaultBudgetAmount, defaultSpendLimit, sortOrder]
      properties:
        id: { type: integer }
        name: { type: string }
        color: { type: string }
        defaultBudgetAmount: { type: number }
        defaultSpendLimit: { type: number }
        sortOrder: { type: integer }
    Transaction:
      type: object
      required: [id, name, amount, type, date, monthKey, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        amount: { type: number }
        type:
          type: string
          enum: [expense, income, transfer]
        categoryId: { type: integer, nullable: true }
        accountId: { type: integer, nullable: true }
        toAccountId: { type: integer, nullable: true }
        date: { type: string, format: date }
        monthKey: { type: string, example: '2026-03' }
        recurringId: { type: integer, nullable: true }
        status:
          type: string
          enum: [confirmed, pending]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AppSettings:
      type: object
      required: [defaultTakeHomePay, currencySymbol, categoryTemplates]
      properties:
        defaultTakeHomePay: { type: number }
        currencySymbol: { type: string, example: '$' }
        categoryTemplates:
          type: array
          items: { $ref: '#/components/schemas/CategoryTemplate' }
...
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code: { type: string }
security:
  - bearerAuth: []
paths:
...
  /api/settings/income:
    put:
      tags: [Settings]
      summary: Update default take-home pay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [defaultTakeHomePay]
              properties:
                defaultTakeHomePay: { type: number }
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppSettings' }
  /api/settings/currency:
    put:
      tags: [Settings]
      summary: Update currency symbol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currencySymbol]
              properties:
                currencySymbol: { type: string, example: '$' }
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppSettings' }
  /api/settings/templates:
    get:
      tags: [Settings]
      summary: List all category templates
      responses:
        '200':
          description: List of category templates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CategoryTemplate' }
    post:
      tags: [Settings]
      summary: Create a category template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, color, defaultBudgetAmount, defaultSpendLimit, sortOrder]
              properties:
                name: { type: string }
                color: { type: string }
                defaultBudgetAmount: { type: number }
                defaultSpendLimit: { type: number }
                sortOrder: { type: integer }
      responses:
        '201':
          description: Created template
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryTemplate' }
    put:
      tags: [Settings]
      summary: Bulk replace all category templates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [name, color, defaultBudgetAmount, defaultSpendLimit, sortOrder]
                properties:
                  name: { type: string }
                  color: { type: string }
                  defaultBudgetAmount: { type: number }
                  defaultSpendLimit: { type: number }
                  sortOrder: { type: integer }
      responses:
        '200':
          description: Replaced templates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CategoryTemplate' }
  /api/settings/templates/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    put:
      tags: [Settings]
      summary: Update a category template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                color: { type: string }
                defaultBudgetAmount: { type: number }
                defaultSpendLimit: { type: number }
                sortOrder: { type: integer }
      responses:
        '200':
          description: Updated template
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryTemplate' }
        '404':
          description: Template not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Settings]
      summary: Delete a category template
      responses:
        '204':
          description: Deleted
        '404':
          description: Template not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ─── Months ────────────────────────────────────────────────────────────────
...
  /api/transactions:
    get:
      tags: [Transactions]
      summary: List transactions with optional filters
      parameters:
        - name: monthKey
          in: query
          schema: { type: string }
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: type
          in: query
          schema:
            type: string
            enum: [expense, income, transfer]
        - name: status
          in: query
          schema:
            type: string
            enum: [confirmed, pending]
        - name: search
          in: query
          schema: { type: string }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Transaction' }
    post:
      tags: [Transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, amount, type, date, monthKey, status]
              properties:
                name: { type: string }
                amount: { type: number }
                type:
                  type: string
                  enum: [expense, income, transfer]
                categoryId: { type: integer, nullable: true }
                accountId: { type: integer, nullable: true }
                toAccountId: { type: integer, nullable: true }
                date: { type: string, format: date }
                monthKey: { type: string }
                recurringId: { type: integer, nullable: true }
                status:
                  type: string
                  enum: [confirmed, pending]
      responses:
        '201':
          description: Created transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
  /api/transactions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    put:
      tags: [Transactions]
      summary: Update a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                amount: { type: number }
                type:
                  type: string
                  enum: [expense, income, transfer]
                categoryId: { type: integer, nullable: true }
                accountId: { type: integer, nullable: true }
                toAccountId: { type: integer, nullable: true }
                date: { type: string, format: date }
                monthKey: { type: string }
                status:
                  type: string
                  enum: [confirmed, pending]
      responses:
        '200':
          description: Updated transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '404':
          description: Transaction not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Transactions]
      summary: Delete a transaction
      responses:
        '204':
          description: Deleted
        '404':
          description: Transaction not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/transactions/{id}/confirm:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Transactions]
      summary: Confirm a pending transaction
      responses:
        '200':
          description: Confirmed transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '404':
          description: Transaction not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/transactions/{id}/dismiss:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    delete:
      tags: [Transactions]
      summary: Dismiss (remove) a pending transaction without confirming
      responses:
        '204':
          description: Dismissed
        '404':
          description: Transaction not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ─── Recurring Transactions ────────────────────────────────────────────────
...